Added seperate test case for cbrt and exp2 functions
